Refactor the mixins
