+++ Client/src/main/java/com/minecrafttas/tasbattle/mixin/gui/MixinRealmsNotificationsScreen.java
@@ -1,4 +1,4 @@
-package com.minecrafttas.tasbattle.mixin;
+package com.minecrafttas.tasbattle.mixin.gui;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
